apps/server/requirements.txt
asgiref==3.7.2
Django==4.2
django-environ==0.11.2
django-rest-framework-supertest==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
